+++ CHANGELOG.md
@@ -4,7 +4,20 @@
 
 ## Unassigned
 
-## 2.0.3 Release
+## 2.0.7
+- [Chore] Update project settings (Xcode 12)
+- [Chore] Update ruby gems
+
+## 2.0.6
+- [Bugfix] Fixes skipping single style attributes after parsing text-decoration attributes (#30)
+
+## 2.0.5
+- [Enhancement] Add parsing completion handler for reporting unknown fonts (#23)
+
+## 2.0.4
+- [Bugfix] Fix a out-of-bounds exception when parsing certain asian characters (#27)
+
+## 2.0.3
 - [Enhancement] Parse `<em>` and `<strong>` tags (#23)
 - [Bugfix] Fix writing and reading of multiple text-decoration attribute values (#22)
 - [Internal] Handle malformed xml more gracefully - early return instead of assert (#21)